pyproject.toml
[project]
name = "squish-mcp"
version = "0.1.0"
description = "Squish MCP demo with context expansion."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.0",
    "psutil>=7.2.2",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
]


[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "ty>=0.0.17",
]

[project.scripts]
squish_mcp = "squish_mcp.server:run"


[tool.ruff]
target-version = "py312"
extend-exclude = [
    "coin",
    "thirdparty",
    "tests/squish/test_data",
]
line-length = 120

[tool.ruff.lint.isort]
case-sensitive = true
lines-after-imports = 2
lines-between-types = 1
force-single-line = true

[tool.ruff.lint]
select = [
    "ANN",  # flake8-annotations (type hints)
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "E",    # pycodestyle
    "F",    # Pyflakes
    "FAST", # FastAPI
    "I",    # isort
    "INT",  # flake8-gettext
    "PL",   # Pylint
    "ARG",  # flake8-unused-arguments
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--ignore=tests/squish/test_data",
]

[tool.uv]
package = true

[tool.ty.rules]
unused-ignore-comment = "warn"
redundant-cast = "ignore"
possibly-missing-attribute = "error"
possibly-missing-import = "error"

[tool.ty.src]
exclude = [
    "tests/squish/test_data",
    "examples",
]


[tool.setuptools.packages.find]
where = ["src"]